AppInstaller // Tearing down AppDelegate into parts.
